refactor(mcp): move wire-name contract to mur-common::mcp_naming
