+++ mur-agent-runtime/src/tools/naming.rs
@@ -1,31 +1,8 @@
-//! MCP wire-name encoding: `mcp__<server>__<tool>`.
-//!
-//! The LLM tool-name field must match `^[a-zA-Z0-9_-]{1,64}$`.
-//! Server names are sanitised by collapsing non-alphanumeric/dash chars into `_`.
+//! MCP wire-name encoding — re-exported from `mur_common::mcp_naming`
+//! (single source of truth; mur-core builds `ToolRule` patterns from the
+//! same contract).
 
-/// Sanitise an MCP server name so it's safe to embed in a wire name.
-///
-/// Collapses any run of non-`[a-zA-Z0-9-]` chars into a single `_`.
-pub fn sanitize_server(name: &str) -> String {
-    let mut out = String::with_capacity(name.len());
-    let mut prev_us = false;
-    for c in name.chars() {
-        if c.is_ascii_alphanumeric() || c == '-' {
-            out.push(c);
-            prev_us = false;
-        } else if !prev_us {
-            out.push('_');
-            prev_us = true;
-        }
-    }
-    // Trim trailing underscore that would make the boundary look odd.
-    out.trim_end_matches('_').to_string()
-}
-
-/// Encode a server + tool name into the `mcp__<server>__<tool>` wire format.
-pub fn wire_name(server_sanitized: &str, tool: &str) -> String {
-    format!("mcp__{server_sanitized}__{tool}")
-}
+pub use mur_common::mcp_naming::{sanitize_server, wire_name};
 
 #[cfg(test)]
 mod tests {
